feat(collector): create collector base code (#1134)
