+++ apps/web/src/services/asset-inventory/collector/create-collector-old/CollectorProgressWizard.vue
